fix: respect env variables in .env for settings.json variable substitution (#3416)
